moving config to json for easier editing
