+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules/**/*
 .expo/**/*
 config.js
+config.json
 credentials.yml
 workspace
 builds
