+++ concourse/scripts/helpers/build_config.sh
@@ -1,9 +1,8 @@
 #!/bin/bash
 
-escaped_url=$(echo "$BOOKIT_API_BASE_URL" | sed 's#\([]\!\(\)\#\%\@\*\$\/&\-\=[]\)#\\\1#g')
-
 main () {
-	sed "s/<bookitApiBaseUrl>/$escaped_url/g" ./config-sample.js > ./config.js;
+	> ./config.json
+	jq '.bookitApiBaseUrl=env.BOOKIT_API_BASE_URL' ./config-sample.json >> ./config.json
 }
 
 main
