--- config-sample.js
@@ -1,6 +1,0 @@
-const config = {
-  bookitApiBaseUrl: '<bookitApiBaseUrl>',
-  facebookAppId: 'get it from facebook!', // https://developers.facebook.com/apps/
-}
-
-export default config
