+++ config-sample.json
@@ -1,0 +1,4 @@
+{
+  "bookitApiBaseUrl": "<bookitApiBaseUrl>",
+  "facebookAppId": "get it from facebook"
+}
