+++ src/api/index.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import config from '../../config'
+import config from '../../config.json'
 
 const baseUrl = config.bookitApiBaseUrl
