+++ src/screens/HomeScreen.js
@@ -3,6 +3,7 @@ import {
   StyleSheet,
   View,
   FlatList,
+  Text,
 } from 'react-native'
 import { connect } from 'react-redux'
 import { DateTime, Duration } from 'luxon'
@@ -20,6 +21,7 @@ import TimePicker from '../components/TimePicker'
 import DurationPicker from '../components/DurationPicker'
 import RootModal from '../components/modals/RootModal'
 import LoginWarning from '../components/LoginWarning'
+import config from '../../config.json'
 
 class HomeScreen extends React.Component {
   static navigationOptions = {
@@ -58,7 +60,12 @@ class HomeScreen extends React.Component {
     const { userExists } = this.props
 
     if (!userExists) {
-      return (
+      return config.nonce ? (
+        <View style={{ marginTop: 90 }}>
+          <Text accessibilityLabel="nonce">{ config.nonce }</Text>
+          <LoginWarning currentScreen="home screen" />
+        </View>
+      ) : (
         <View style={{ marginTop: 90 }}>
           <LoginWarning currentScreen="home screen" />
         </View>
